+++ example/lib/web_workers/fibonacci.dart
@@ -1,9 +1,0 @@
-import 'package:isolate_manager/isolate_manager.dart';
-
-import '../functions.dart';
-
-/// dart compile js fibonacci.dart -o fibonacci.js -O4
-
-main() {
-  IsolateManagerFunction.workerFunction(fibonacci);
-}