+++ lib/isolate_manager.dart
@@ -5,3 +5,4 @@ export 'src/base/isolate_contactor.dart'
 export 'src/isolate_manager.dart';
 export 'src/isolate_manager_controller.dart';
 export 'src/isolate_manager_function.dart';
+export 'src/models/isolate_manager_worker.dart';